// payload-types.ts
export interface Config {
  auth: {
    users: UserAuthOperations;
  };
  blocks: {
    button: Button;
    carousel: Carousel;
    'contact-form': ContactForm;
    gallery: Gallery;
    headline: Headline;
    'status-alert': StatusAlert;
    text: Text;
    section: Section;
  };
  collections: {
    pages: Page;
    media: Media;
    users: User;
    dialogs: Dialog;
    'payload-kv': PayloadKv;
    'payload-locked-documents': PayloadLockedDocument;
    'payload-preferences': PayloadPreference;
    'payload-migrations': PayloadMigration;
  };
  collectionsJoins: {};
  collectionsSelect: {
    pages: PagesSelect<false> | PagesSelect<true>;
    media: MediaSelect<false> | MediaSelect<true>;
    users: UsersSelect<false> | UsersSelect<true>;
    dialogs: DialogsSelect<false> | DialogsSelect<true>;
    'payload-kv': PayloadKvSelect<false> | PayloadKvSelect<true>;
    'payload-locked-documents': PayloadLockedDocumentsSelect<false> | PayloadLockedDocumentsSelect<true>;
    'payload-preferences': PayloadPreferencesSelect<false> | PayloadPreferencesSelect<true>;
    'payload-migrations': PayloadMigrationsSelect<false> | PayloadMigrationsSelect<true>;
  };
  db: {
    defaultIDType: number;
  };
  fallbackLocale: null;
  globals: {
    settings: Setting;
    downloads: Download;
    navigation: Navigation;
  };
  globalsSelect: {
    settings: SettingsSelect<false> | SettingsSelect<true>;
    downloads: DownloadsSelect<false> | DownloadsSelect<true>;
    navigation: NavigationSelect<false> | NavigationSelect<true>;
  };
  locale: null;
  user: User & {
    collection: 'users';
  };
  jobs: {
    tasks: unknown;
    workflows: unknown;
  };
}
export interface UserAuthOperations {
  forgotPassword: {
    email: string;
    password: string;
  };
  login: {
    email: string;
    password: string;
  };
  registerFirstUser: {
    email: string;
    password: string;
  };
  unlock: {
    email: string;
    password: string;
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "button".
 */
export interface Button {
  label: string;
  type: 'page' | 'custom' | 'dialog' | 'file';
  page?: (number | null) | Page;
  file?: (number | null) | Media;
  url?: string | null;
  /**
   * If set, pressing the button will open the selected dialog.
   */
  dialog?: (number | null) | Dialog;
  id?: string | null;
  blockName?: string | null;
  blockType: 'button';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: number;
  title: string;
  slug: string;
  keywords?: string | null;
  description?: string | null;
  content?:
    | {
        /**
         * Allows to create a links to the section
         */
        slug?: string | null;
        children?:
          | (
              | {
                  label: string;
                  type: 'page' | 'custom' | 'dialog' | 'file';
                  page?: (number | null) | Page;
                  file?: (number | null) | Media;
                  url?: string | null;
                  /**
                   * If set, pressing the button will open the selected dialog.
                   */
                  dialog?: (number | null) | Dialog;
                  id?: string | null;
                  blockName?: string | null;
                  blockType: 'button';
                }
              | {
                  text?: string | null;
                  images?: (number | Media)[] | null;
                  id?: string | null;
                  blockName?: string | null;
                  blockType: 'carousel';
                }
              | {
                  lines: {
                    text: string;
                    id?: string | null;
                  }[];
                  id?: string | null;
                  blockName?: string | null;
                  blockType: 'headline';
                }
              | {
                  color: string;
                  content: string;
                  id?: string | null;
                  blockName?: string | null;
                  blockType: 'status-alert';
                }
              | {
                  variant?: ('p1' | 'p2' | 'p3') | null;
                  text?: string | null;
                  id?: string | null;
                  blockName?: string | null;
                  blockType: 'text';
                }
              | {
                  id?: string | null;
                  blockName?: string | null;
                  blockType: 'contact-form';
                }
              | {
                  title?: string | null;
                  items?:
                    | {
                        images: (number | Media)[];
                        title: string;
                        category: 'kilns' | 'controllers' | 'accessories' | 'other';
                        /**
                         * e.g. '$100', 'from $50'
                         */
                        price?: string | null;
                        description?: string | null;
                        id?: string | null;
                      }[]
                    | null;
                  id?: string | null;
                  blockName?: string | null;
                  blockType: 'gallery';
                }
              | {
                  triggerLabel: string;
                  slides?:
                    | {
                        image: number | Media;
                        description?: string | null;
                        id?: string | null;
                      }[]
                    | null;
                  id?: string | null;
                  blockName?: string | null;
                  blockType: 'realizations';
                }
            )[]
          | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'section';
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "dialogs".
 */
export interface Dialog {
  id: number;
  title?: string | null;
  description?: string | null;
  content?:
    | {
        id?: string | null;
        blockName?: string | null;
        blockType: 'contact-form';
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "carousel".
 */
export interface Carousel {
  text?: string | null;
  images?: (number | Media)[] | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'carousel';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "contact-form".
 */
export interface ContactForm {
  id?: string | null;
  blockName?: string | null;
  blockType: 'contact-form';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "gallery".
 */
export interface Gallery {
  title?: string | null;
  items?:
    | {
        images: (number | Media)[];
        title: string;
        category: 'kilns' | 'controllers' | 'accessories' | 'other';
        /**
         * e.g. '$100', 'from $50'
         */
        price?: string | null;
        description?: string | null;
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'gallery';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "headline".
 */
export interface Headline {
  lines: {
    text: string;
    id?: string | null;
  }[];
  id?: string | null;
  blockName?: string | null;
  blockType: 'headline';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "status-alert".
 */
export interface StatusAlert {
  color: string;
  content: string;
  id?: string | null;
  blockName?: string | null;
  blockType: 'status-alert';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "text".
 */
export interface Text {
  variant?: ('p1' | 'p2' | 'p3') | null;
  text?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'text';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "section".
 */
export interface Section {
  /**
   * Allows to create a links to the section
   */
  slug?: string | null;
  children?:
    | (
        | {
            label: string;
            type: 'page' | 'custom' | 'dialog' | 'file';
            page?: (number | null) | Page;
            file?: (number | null) | Media;
            url?: string | null;
            /**
             * If set, pressing the button will open the selected dialog.
             */
            dialog?: (number | null) | Dialog;
            id?: string | null;
            blockName?: string | null;
            blockType: 'button';
          }
        | {
            text?: string | null;
            images?: (number | Media)[] | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'carousel';
          }
        | {
            lines: {
              text: string;
              id?: string | null;
            }[];
            id?: string | null;
            blockName?: string | null;
            blockType: 'headline';
          }
        | {
            color: string;
            content: string;
            id?: string | null;
            blockName?: string | null;
            blockType: 'status-alert';
          }
        | {
            variant?: ('p1' | 'p2' | 'p3') | null;
            text?: string | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'text';
          }
        | {
            id?: string | null;
            blockName?: string | null;
            blockType: 'contact-form';
          }
        | {
            title?: string | null;
            items?:
              | {
                  images: (number | Media)[];
                  title: string;
                  category: 'kilns' | 'controllers' | 'accessories' | 'other';
                  /**
                   * e.g. '$100', 'from $50'
                   */
                  price?: string | null;
                  description?: string | null;
                  id?: string | null;
                }[]
              | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'gallery';
          }
        | {
            triggerLabel: string;
            slides?:
              | {
                  image: number | Media;
                  description?: string | null;
                  id?: string | null;
                }[]
              | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'realizations';
          }
      )[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'section';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
  id: number;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "payload-kv".
 */
export interface PayloadKv {
  id: number;
  key: string;
  data:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
}
// ...
export interface PayloadPreference {
  id: number;
  user: {
    relationTo: 'users';
    value: number | User;
  };
  key?: string | null;
  value?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "payload-migrations".
 */
export interface PayloadMigration {
  id: number;
  name?: string | null;
  batch?: number | null;
  updatedAt: string;
  createdAt: string;
}